+++ tests/oihana/files/GetMimeTypeTest.php
@@ -1,10 +1,13 @@
 <?php
 
-namespace oihana\files ;
+namespace tests\oihana\files ;
 
 use PHPUnit\Framework\TestCase;
 
 use oihana\files\exceptions\DirectoryException;
+
+use function oihana\files\deleteDirectory;
+use function oihana\files\getMimeType;
 
 class GetMimeTypeTest extends TestCase
 {